Extended TI Basic Module  * Added STEP part to FOR-TO-STEP statement  * Implemented fractional FOR-TO-STEP initializer and step
Extended testing
 * Added two new examples from User's Reference Guide

+++ src/main/kotlin/com/github/mmrsic/ti99/basic/TiBasicInterpreter.kt
@@ -109,7 +109,7 @@ class TiBasicProgramInterpreter(machine: TiBasicModule, private val codeSequence
     }
 
     /** Begin a new for-loop. */
-    fun beginForLoop(lineNumber: Int, stmt: LetNumberStatement, limit: NumericExpr, increment: Int = 1) {
+    fun beginForLoop(lineNumber: Int, stmt: LetNumberStatement, limit: NumericExpr, increment: Double = 1.0) {
         val program = machine.program ?: throw IllegalArgumentException("Can't begin for-loop without program")
         val jumpLineNumber = program.nextLineNumber(lineNumber)
             ?: throw IllegalArgumentException("Line number has no successor: $lineNumber")
